pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yoda-must"
version = "0.9.0"
description = "A CLI tool to build, push, and deploy services to a VPS"
authors = [
  { name="x0y", email="x0y@milkyweb.co" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

dependencies = [
  "python-dotenv>=1.1.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
must = "yoda.cli:main"
